11-14-2015 Changes
officers.

This version contains fixes for the incident map and also use cluster
items for the map

Contains a Render class to render the criminal icons and cluster image

Contains the ability to get directions to the officer’s reported
incident.

Contains the ability to see the officer’s incident report in text view

Contains the ability to contact user

Fixed the bug where latitude and longitude were stored in the wrong
fields in tbl_incidents

Fixed the issue where xml byte exception was thrown when loading the
incident map.

Fixed the error when attempting to update the
tbl_reponds_officer_incident when responding to an incident